ui/requirements.txt
streamlit==1.4.0
st-annotated-text==2.0.0
